Refactor rotary encoder handling and improve frequency tuning logic
- Updated the ValueChangeDialog to handle rotary encoder values more effectively, allowing for direct adjustments based on encoder input.
- Enhanced frequency tuning in AmDisplay by refining the handling of rotary encoder states and ensuring proper frequency limits are enforced.
- Modified the Band structure to use int16_t for lastBFO to accommodate a wider range of values.
- Improved the display logic in SevenSegmentFreq to correctly format and display frequencies based on the current mode (AM, FM, SSB, CW).
- Adjusted the rotary encoder configuration to allow for variable steps per notch, enhancing user experience during frequency adjustments.
- Cleaned up code in various files for better readability and maintainability, including removing unnecessary comments and improving variable naming.
- Fixed potential bugs related to frequency stepping and BFO adjustments in both AM and FM modes.

+++ Band.h
@@ -37,8 +37,8 @@ struct BandTableVar {
     uint8_t currStep;     // Aktuális lépésköz (növelés/csökkentés)
     uint8_t currMod;      // Aktuális mód/modulációs típus (FM, AM, LSB, USB, CW)
     uint16_t antCap;      // Antenna Tuning Capacitor
-    uint8_t lastBFO;      // Utolsó BFO érték
-    uint8_t lastmanuBFO;  // Utolsó manuális BFO érték X-Tal segítségével
+    int16_t lastBFO;      // Utolsó BFO érték
+    int16_t lastmanuBFO;  // Utolsó manuális BFO érték X-Tal segítségével
 };
 
 // A kombinált struktúra az állandó és változó adatok összekapcsolásával
